+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-This changelog is for the main miru package.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -9,3 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Changelog
+- `list` subcommand
+- `increment` subcommand
+- `login` option
+- `no-cache` option